bugfix: split between ApiService and Api endpoint was rather wrong
