+++ src/main/scala/hatdex/hat/api/ApiService.scala
@@ -11,16 +11,47 @@ package hatdex.hat.api
 
 import akka.actor.{ActorLogging, ActorRefFactory}
 import akka.event.LoggingAdapter
+import hatdex.hat.api.endpoints._
 import spray.routing._
+import spray.util.LoggingContext
 
 // we don't implement our route structure directly in the service actor because
 // we want to be able to test it independently, without having to spin up an actor
 class ApiService extends HttpServiceActor with ActorLogging with Cors {
-  val api = new Api {
-    override val logger: LoggingAdapter = log
-    override implicit def actorRefFactory: ActorRefFactory = context
+  val apiLogger = LoggingContext.fromActorRefFactory
+
+  // The HttpService trait defines only one abstract member, which
+  // connects the services environment to the enclosing actor or test.
+  // We also want logging, hence the abstract logger member is included
+  // in HAT services as well
+
+  trait LoggingHttpService {
+    def actorRefFactory = context
+    val logger = log
   }
 
-  def receive = runRoute(api.routes)
+  val api = new Api {
+    implicit def actorRefFactory: ActorRefFactory = context
+
+    // Initialise all the service the actor handles
+    val helloService = new Hello with LoggingHttpService
+    val apiDataService = new Data with LoggingHttpService
+    val apiBundleService = new Bundles with LoggingHttpService
+    val dataDebitService = new DataDebit with LoggingHttpService
+    val apiPropertyService = new Property with LoggingHttpService
+    val eventsService = new Event with LoggingHttpService
+    val locationsService = new Location with LoggingHttpService
+    val peopleService = new Person with LoggingHttpService
+    val thingsService = new Thing with LoggingHttpService
+    val organisationsService = new Organisation with LoggingHttpService
+    val userService = new Users with LoggingHttpService
+    val typeService = new Type with LoggingHttpService
+  }
+
+  val routes = logRequestResponse(api.requestMethodAndResponseStatusAsInfo _) {
+    api.routes
+  }
+
+  def receive = runRoute(routes)
 }
